requirements.txt
streamlit>=1.43.0
langchain>=0.3.21
langchain-openai>=0.3.8
langchain-ollama>=0.3.3
langchain-google-genai>=2.1.5
python-dotenv>=1.0.1
pydantic>=2.10.6
fastapi>=0.115.12
uvicorn>=0.34.0
requests>=2.32.3
PyJWT[crypto]>=2.10.1
pytest>=8.3.5
